Make installer fall back to hard links in case symbolic link creation fails.

+++ scripts/inc.installer.sh
@@ -87,7 +87,7 @@ for f in '"$CHMODXFILES"'; do
 done
 install -d "$l"
 for i in $($bb --list); do
-  if ! ln -sf "$bb" "$l/$i" && ! $bb ln -sf "$bb" "$l/$i" ; then
+  if ! ln -sf "$bb" "$l/$i" && ! $bb ln -sf "$bb" "$l/$i" && ! $bb ln -f "$bb" "$l/$i" ; then
     echo "ui_print ERROR 10: Failed to set-up Open GApps'"'"' pre-bundled '"$2"' binary" > "$OUTFD"
     echo "ui_print" > "$OUTFD"
     echo "ui_print Please use TWRP as recovery instead" > "$OUTFD"
